# payments_openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/create_account":{"post":{"summary":"Create Account","operationId":"create_account_create_account_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payment"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/topup":{"post":{"summary":"Topup","operationId":"topup_topup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payment"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/balance/{user_id}":{"get":{"summary":"Get Balance","operationId":"get_balance_balance__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/credit":{"post":{"summary":"Credit","operationId":"credit_credit_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/withdraw":{"post":{"summary":"Withdraw","operationId":"withdraw_withdraw_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payment"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Payment":{"properties":{"user_id":{"type":"integer","title":"User Id"},"amount":{"type":"integer","title":"Amount"}},"type":"object","required":["user_id","amount"],"title":"Payment"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}